refactor: place order_graph_validator in client directory
